// src/app/model/document.ts
export interface Document {
    DId?: number;
    DartaNo?: number;
    DocumentTitle?: string;
    Organization?: string;
    StateId?: number;
    DistrictId?: number;
    VDCMUNId?: number;
    WardNo?: number;
    Tole?: string;
    Email?: string;
    Phone?: string;
    MobileNo?: string;
    EntryDate?: Date;
    UserId?: number;
    DocumentPath?: string;
    IsActive?: boolean;
    Remarks?: string;
  }
  